add object detection functionality
