--- DesignScape/DesignScape/Extension/UIColor+Hex.swift
@@ -1,34 +1,0 @@
-//
-//  UIColor+Hex.swift
-//  DesignScape
-//
-//  Created by Minh Huynh on 2/24/24.
-//
-
-import UIKit
-
-extension UIColor {
-    private static func intFromHexString(hexStr: String) -> UInt32 {
-        var hexInt: UInt32 = 0
-        // Create scanner
-        let scanner: Scanner = Scanner(string: hexStr)
-        // Tell scanner to skip the # character
-        scanner.charactersToBeSkipped = CharacterSet(charactersIn: "#")
-        // Scan hex value
-        hexInt = UInt32(bitPattern: scanner.scanInt32(representation: .hexadecimal) ?? 0)
-        return hexInt
-    }
-
-    static func colorFromHexString(hexString: String, alpha: CGFloat = 1.0) -> UIColor {
-        // Convert hex string to an integer
-        let hexint = Int(intFromHexString(hexStr: hexString))
-        let red = CGFloat((hexint & 0xff0000) >> 16) / 255.0
-        let green = CGFloat((hexint & 0xff00) >> 8) / 255.0
-        let blue = CGFloat((hexint & 0xff) >> 0) / 255.0
-
-        // Create color object, specifying alpha as well
-        let color = UIColor(red: red, green: green, blue: blue, alpha: alpha)
-        return color
-    }
-}
-
